+++ Abc.Future/MaybeEx.Async.cs
@@ -1,0 +1,74 @@
+﻿// See LICENSE.txt in the project root for license information.
+
+namespace Abc
+{
+    using System;
+    using System.Diagnostics.Contracts;
+    using System.Threading.Tasks;
+
+    using Anexn = System.ArgumentNullException;
+
+    // Async methods.
+    public partial class MaybeEx
+    {
+        // Configurable core async methods?
+        // https://devblogs.microsoft.com/dotnet/configureawait-faq/
+        // https://ericlippert.com/2020/03/10/passing-awaited-tasks/
+
+        [Pure]
+        public static async Task<Maybe<TResult>> BindAsync<T, TResult>(
+            Maybe<T> maybe,
+            Func<T, Task<Maybe<TResult>>> binder,
+            bool continueOnCapturedContext)
+        {
+            if (binder is null) { throw new Anexn(nameof(binder)); }
+
+            if (maybe.TryGetValue(out T value))
+            {
+                return await binder(value)
+                    .ConfigureAwait(continueOnCapturedContext);
+            }
+            else
+            {
+                return Maybe<TResult>.None;
+            }
+        }
+
+        [Pure]
+        public static async Task<Maybe<TResult>> SelectAsync<T, TResult>(
+            Maybe<T> maybe,
+            Func<T, Task<TResult>> selector,
+            bool continueOnCapturedContext)
+        {
+            if (selector is null) { throw new Anexn(nameof(selector)); }
+
+            if (maybe.TryGetValue(out T value))
+            {
+                TResult result = await selector(value)
+                    .ConfigureAwait(continueOnCapturedContext);
+                return Maybe.Of(result);
+            }
+            else
+            {
+                return Maybe<TResult>.None;
+            }
+        }
+
+        //
+        // More async methods.
+        //
+
+        [Pure]
+        public static async Task<Maybe<T>> WhereAsync<T>(
+            Maybe<T> maybe,
+            Func<T, Task<bool>> predicate)
+        {
+            if (predicate is null) { throw new Anexn(nameof(predicate)); }
+
+            return await maybe.BindAsync(__binder).ConfigureAwait(false);
+
+            async Task<Maybe<T>> __binder(T x)
+                => await predicate(x).ConfigureAwait(false) ? maybe : Maybe<T>.None;
+        }
+    }
+}