+++ Abc.Tests/MaybeTests.cs
@@ -13,6 +13,9 @@ namespace Abc
 
     public static partial class MaybeTests
     {
+        public const string NullString = null;
+        public const string? NullNullString = null;
+
         private static readonly Maybe<int> Ø = Maybe<int>.None;
         private static readonly Maybe<long> ØL = Maybe<long>.None;
 
@@ -986,112 +989,6 @@ namespace Abc
         #endregion
     }
 
-    // "Bitwise" logical operations.
-    public partial class MaybeTests
-    {
-        [Fact]
-        public static void OrElse()
-        {
-            // Some Some -> Some
-            Assert.Equal(One, One.OrElse(Two));
-            // Some None -> Some
-            Assert.Equal(One, One.OrElse(Ø));
-            // None Some -> Some
-            Assert.Equal(Two, Ø.OrElse(Two));
-            // None None -> None
-            Assert.Equal(Ø, Ø.OrElse(Ø));
-        }
-
-        [Fact]
-        public static void AndThen()
-        {
-            // Some Some -> Some
-            Assert.Equal(TwoL, One.AndThen(TwoL));
-            // Some None -> None
-            Assert.Equal(ØL, One.AndThen(ØL));
-            // None Some -> None
-            Assert.Equal(ØL, Ø.AndThen(TwoL));
-            // None None -> None
-            Assert.Equal(ØL, Ø.AndThen(ØL));
-        }
-
-        [Fact]
-        public static void XorElse()
-        {
-            // Some Some -> None
-            Assert.Equal(Ø, One.XorElse(Two));
-            // Some None -> Some
-            Assert.Equal(One, One.XorElse(Ø));
-            // None Some -> Some
-            Assert.Equal(Two, Ø.XorElse(Two));
-            // None None -> None
-            Assert.Equal(Ø, Ø.XorElse(Ø));
-
-            // XorElse() flips to itself.
-            Assert.Equal(Ø, Two.XorElse(One));
-            Assert.Equal(One, Ø.XorElse(One));
-            Assert.Equal(Two, Two.XorElse(Ø));
-            Assert.Equal(Ø, Ø.XorElse(Ø));
-        }
-
-        [Fact]
-        public static void BitwiseOr()
-        {
-            // Some Some -> Some
-            Assert.Equal(One, One | Two);
-            Assert.Equal(Two, Two | One);   // non-abelian
-            // Some None -> Some
-            Assert.Equal(One, One | Ø);
-            // None Some -> Some
-            Assert.Equal(Two, Ø | Two);
-            // None None -> None
-            Assert.Equal(Ø, Ø | Ø);
-
-            Assert.LogicalTrue(One | Two);
-            Assert.LogicalTrue(One | Ø);
-            Assert.LogicalTrue(Ø | Two);
-            Assert.LogicalFalse(Ø | Ø);
-        }
-
-        [Fact]
-        public static void BitwiseAnd()
-        {
-            // Some Some -> Some
-            Assert.Equal(Two, One & Two);
-            Assert.Equal(One, Two & One);   // non-abelian
-            // Some None -> None
-            Assert.Equal(Ø, One & Ø);
-            // None Some -> None
-            Assert.Equal(Ø, Ø & Two);
-            // None None -> None
-            Assert.Equal(Ø, Ø & Ø);
-
-            Assert.LogicalTrue(One & Two);
-            Assert.LogicalFalse(One & Ø);
-            Assert.LogicalFalse(Ø & Two);
-            Assert.LogicalFalse(Ø & Ø);
-        }
-
-        [Fact]
-        public static void ExclusiveOr()
-        {
-            // Some Some -> None
-            Assert.Equal(Ø, One ^ Two);
-            Assert.Equal(Ø, Two ^ One);     // abelian
-            // Some None -> Some
-            Assert.Equal(One, One ^ Ø);
-            // None Some -> Some
-            Assert.Equal(Two, Ø ^ Two);
-            // None None -> None
-            Assert.Equal(Ø, Ø ^ Ø);
-
-            Assert.LogicalFalse(One ^ Two);
-            Assert.LogicalTrue(One ^ Ø);
-            Assert.LogicalTrue(Ø ^ Two);
-            Assert.LogicalFalse(Ø ^ Ø);
-        }
-    }
-
     // Misc methods.
     public partial class MaybeTests
     {
@@ -1260,6 +1157,63 @@ namespace Abc
             Assert.True(Maybe.SomeOrNone("XXX").Contains("XXX", StringComparer.OrdinalIgnoreCase));
             Assert.True(Maybe.SomeOrNone("XXX").Contains("xxx", StringComparer.OrdinalIgnoreCase));
             Assert.False(Maybe.SomeOrNone("XXX").Contains("yyy", StringComparer.OrdinalIgnoreCase));
+        }
+    }
+
+    // Not actually a test of Maybe.
+    public partial class MaybeTests
+    {
+        // Comparison w/ null is weird.
+        //   "For the comparison operators <, >, <=, and >=, if one or both
+        //   operands are null, the result is false; otherwise, the contained
+        //   values of operands are compared. Do not assume that because a
+        //   particular comparison (for example, <=) returns false, the opposite
+        //   comparison (>) returns true."
+        // https://docs.microsoft.com/en-us/dotnet/csharp/language-reference/builtin-types/nullable-value-types#lifted-operators
+        //   "Basically the rule here is that nulls compare for equality normally,
+        //   but any other comparison results in false."
+        // https://ericlippert.com/2015/08/31/nullable-comparisons-are-weird/
+        // Three options (extracts from Lippert's article):
+        // 1) Make comparison operators produce nullable bool.
+        // 2) Make comparison operators produce bool, and say that
+        //    greater-than-or-equal comparisons to null have the same semantics
+        //    as "or-ing" together the greater-than and equals operations.
+        // 3) Make comparison operators produce a bool and apply a total
+        //    ordering.
+        // Choice 3 is for sorting.
+        [Fact]
+        public static void Comparisons()
+        {
+            int? one = 1;
+            int? nil = null;
+            // Default comparer for nullable int.
+            var comparer = Comparer<int?>.Default;
+
+            // If one of the operand is null, the comparison returns false.
+            // Important consequence: we can't say that "x >= y" is equivalent
+            // to "not(x < y)"...
+            Assert.False(one < nil);    // false
+            Assert.False(one > nil);    // false    "contradicts" Compare; see below
+            Assert.False(one <= nil);   // false
+            Assert.False(one >= nil);   // false
+
+            // Equality is fine.
+            Assert.False(one == nil);   // false
+            Assert.True(one != nil);    // true
+
+#pragma warning disable CS1718 // Comparison made to same variable
+            Assert.False(nil < nil);    // false
+            Assert.False(nil > nil);    // false
+            Assert.False(nil <= nil);   // false    weird
+            Assert.False(nil >= nil);   // false    weird
+
+            Assert.True(nil == nil);    // true
+            Assert.False(nil != nil);   // false
+#pragma warning restore CS1718
+
+            Assert.Equal(1, comparer.Compare(one, nil));    // "one > nil"
+            Assert.Equal(-1, comparer.Compare(nil, one));   // "nil < one"
+            Assert.Equal(0, comparer.Compare(nil, nil));    // "nil >= nil"
         }
     }